+++ wazo_confd/plugins/pjsip/notifier.py
@@ -4,6 +4,9 @@
 from xivo_bus.resources.pjsip.event import (
     PJSIPGlobalUpdatedEvent,
     PJSIPSystemUpdatedEvent,
+    CreateSIPTransportEvent,
+    EditSIPTransportEvent,
+    DeleteSIPTransportEvent,
 )
 
 from wazo_confd import bus, sysconfd
@@ -34,3 +37,39 @@ class PJSIPConfigurationNotifier:
 
 def build_notifier():
     return PJSIPConfigurationNotifier(bus, sysconfd)
+
+
+class PJSIPTransportNotifier:
+    def __init__(self, bus, sysconfd, schema):
+        self.bus = bus
+        self.sysconfd = sysconfd
+        self.schema = schema
+
+    def send_sysconfd_handlers(self, ipbx_command):
+        handlers = {'ipbx': [ipbx_command]}
+        self.sysconfd.exec_request_handlers(handlers)
+
+    def created(self, transport):
+        ipbx_command = 'module reload res_pjsip.so'
+        self.send_sysconfd_handlers(ipbx_command)
+        body = self.schema.dump(transport)
+        event = CreateSIPTransportEvent(body)
+        self.bus.send_bus_event(event)
+
+    def deleted(self, transport):
+        ipbx_command = 'module reload res_pjsip.so'
+        self.send_sysconfd_handlers(ipbx_command)
+        body = self.schema.dump(transport)
+        event = DeleteSIPTransportEvent(body)
+        self.bus.send_bus_event(event)
+
+    def edited(self, transport):
+        ipbx_command = 'module reload res_pjsip.so'
+        self.send_sysconfd_handlers(ipbx_command)
+        body = self.schema.dump(transport)
+        event = EditSIPTransportEvent(body)
+        self.bus.send_bus_event(event)
+
+
+def build_pjsip_transport_notifier(schema):
+    return PJSIPTransportNotifier(bus, sysconfd, schema)